Refactor validation test cases
